Moved SqlTemplate to core package
