+++ sqltemplate-example/src/main/java/springboot/sqltemplate/example/Application.java
@@ -3,13 +3,23 @@ package springboot.sqltemplate.example;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import springboot.sqltemplate.core.SqlTemplate;
 
 @Configuration
 @ComponentScan
 @EnableAutoConfiguration
 public class Application {
+
+    @Bean
+    SqlTemplate sqlTemplate(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+        return new SqlTemplate(jdbcTemplate, namedParameterJdbcTemplate);
+    }
+
     public static void main(String[] args) {
         ConfigurableApplicationContext context = SpringApplication.run(Application.class, args);
         SampleProcess sampleProcess = context.getBean(SampleProcess.class);
